# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Plataforma DTI Níjar — API
  version: "1.0.0"
  description: |
    API REST/JSON de la **Plataforma de Destino Turístico Inteligente (DTI) Smart City**
    del Ayuntamiento de Níjar.

    **Expediente:** 18962/2025
    **Adjudicatario:** IT DIGITTAL
    **Marco:** PRTR — NextGenerationEU — Componente 14
    **Estándares:** UNE 178104 · FIWARE Smart Data Models · ENS Nivel Medio · WCAG 2.1 AA

    ## Versiones y compatibilidad
    - Esta especificación corresponde a la **API v1**.
    - Las rutas están prefijadas con `/api/v1`.
    - Compatibilidad retroactiva garantizada en versiones menores.

    ## Autenticación
    Todos los endpoints (excepto `health`, `ready` y los de autenticación) requieren un
    token JWT válido emitido por `/api/v1/auth/login`. Los tokens incluyen scopes RBAC
    asociados al rol del usuario (administrador_tic, gestor_contenidos, analista_datos,
    operador_smart_office, auditor).
  contact:
    name: IT DIGITTAL
    url: https://www.turismonijar.es
  license:
    name: Propiedad del Ayuntamiento de Níjar (PCAP Anexo XIII)
servers:
  - url: https://api.dti.nijar.es/api/v1
    description: Producción
  - url: https://staging.api.dti.nijar.es/api/v1
    description: Pre-producción
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Iniciar sesión
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Tokens emitidos
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refrescar el token de acceso
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: Nuevo par de tokens
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /data/iot/ingest:
    post:
      tags: [iot]
      summary: Ingerir una observación IoT (HTTP)
      description: |
        Endpoint alternativo al broker MQTT para sistemas que solo pueden publicar por HTTP.
        Aplica las mismas validaciones (rango, deduplicación, normalización semántica).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ObservacionIn" }
      responses:
        "202":
          description: Aceptada para procesamiento
          content:
            application/json:
              schema:
                type: object
                properties:
                  observacion_id: { type: string, format: uuid }
  /data/iot/sensors:
    get:
      tags: [iot]
      summary: Catálogo de sensores
      parameters:
        - { in: query, name: tipo, schema: { type: string } }
        - { in: query, name: estado, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: page_size, schema: { type: integer, default: 20, maximum: 200 } }
      responses:
        "200":
          description: Lista paginada de sensores
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedSensores" }
  /data/iot/sensors/{sensor_id}/observations:
    parameters:
      - { in: path, name: sensor_id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [iot]
      summary: Histórico de observaciones de un sensor
      parameters:
        - { in: query, name: desde, schema: { type: string, format: date-time } }
        - { in: query, name: hasta, schema: { type: string, format: date-time } }
        - { in: query, name: limit, schema: { type: integer, default: 100, maximum: 10000 } }
      responses:
        "200":
          description: Lista de observaciones
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ObservacionOut" }

  # ---------------- SOCIAL LISTENING ----------------
  /data/social/mentions:
    get:
      tags: [social-listening]
      summary: Menciones del destino capturadas en RRSS
      parameters:
        - { in: query, name: fuente, schema: { type: string } }
        - { in: query, name: sentimiento, schema: { type: string, enum: [positivo, neutro, negativo] } }
        - { in: query, name: desde, schema: { type: string, format: date-time } }
        - { in: query, name: hasta, schema: { type: string, format: date-time } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: page_size, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        "200":
          description: Lista paginada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedOpiniones" }
  /data/social/kpis/sentiment:
    get:
      tags: [social-listening]
      summary: KPI de sentimiento medio del destino
      parameters:
        - { in: query, name: desde, schema: { type: string, format: date-time } }
        - { in: query, name: hasta, schema: { type: string, format: date-time } }
        - { in: query, name: granularidad, schema: { type: string, enum: [hora, dia, semana, mes], default: dia } }
      responses:
        "200":
          description: Serie temporal de sentimiento
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SentimentSeries" }
...
  /data/social/topics:
    get:
      tags: [social-listening]
      summary: Temas más mencionados
      parameters:
        - { in: query, name: desde, schema: { type: string, format: date-time } }
        - { in: query, name: hasta, schema: { type: string, format: date-time } }
        - { in: query, name: limit, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: Top temas
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    tema: { type: string }
                    menciones: { type: integer }
                    sentimiento_medio: { type: number }

  # ---------------- CMS ----------------
...
  /chatbot/query:
    post:
      tags: [chatbot]
      summary: Consultar al asistente de turismo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChatQueryIn" }
      responses:
        "200":
          description: Respuesta del chatbot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatResponseOut" }
  /chatbot/feedback:
    post:
      tags: [chatbot]
      summary: Registrar feedback de una interacción
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sesion_id, interaccion_id, util]
              properties:
                sesion_id: { type: string }
                interaccion_id: { type: string }
                util: { type: boolean }
                comentario: { type: string }
      responses:
        "204": { description: Feedback registrado }
...
  /dashboards/smart-office/environment:
    get:
      tags: [dashboards]
      summary: Indicadores ambientales (CO₂, temperatura, humedad, ruido)
      parameters:
        - { in: query, name: desde, schema: { type: string, format: date-time } }
        - { in: query, name: hasta, schema: { type: string, format: date-time } }
        - { in: query, name: granularidad, schema: { type: string, enum: [minuto, hora, dia], default: hora } }
      responses:
        "200":
          description: Series temporales
          content:
            application/json:
              schema: { type: object }
...
  /dashboards/totems/usage:
    get:
      tags: [dashboards]
      summary: Estadísticas de uso de los tótems
      parameters:
        - { in: query, name: desde, schema: { type: string, format: date-time } }
        - { in: query, name: hasta, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: Métricas de uso
          content:
            application/json:
              schema: { type: object }
  /dashboards/reports/monthly:
    get:
      tags: [dashboards]
      summary: Datos del informe mensual de servicio (C.1)
      parameters:
        - { in: query, name: year, required: true, schema: { type: integer } }
        - { in: query, name: month, required: true, schema: { type: integer, minimum: 1, maximum: 12 } }
      responses:
        "200":
          description: Datos del informe mensual
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MonthlyReport" }

# ===================== COMPONENTES =====================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Credenciales inválidas o token expirado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object }
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    TokenResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: bearer }
        expires_in: { type: integer }
    GeoPoint:
      type: object
      required: [type, coordinates]
      properties:
        type: { type: string, enum: [Point] }
        coordinates:
          type: array
          items: { type: number }
          minItems: 2
          maxItems: 2
...
    SensorOut:
      type: object
      properties:
        id: { type: string, format: uuid }
        urn: { type: string }
        nombre: { type: string }
        tipo: { type: string }
        estado: { type: string }
        ubicacion: { $ref: "#/components/schemas/GeoPoint" }
    PaginatedSensores:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SensorOut" }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
    ObservacionIn:
      type: object
      required: [sensor_urn, observado_en]
      properties:
        sensor_urn: { type: string }
        observado_en: { type: string, format: date-time }
        valor: { type: number }
        valores: { type: object }
        unidades: { type: string }
        payload_original: { type: object }
    ObservacionOut:
      type: object
      properties:
        id: { type: string, format: uuid }
        sensor_id: { type: string, format: uuid }
        observado_en: { type: string, format: date-time }
        valor: { type: number }
        valores: { type: object }
        unidades: { type: string }
        valido: { type: boolean }

    # ---- Opinion / Social ----
    OpinionOut:
      type: object
      properties:
        id: { type: string, format: uuid }
        fuente: { type: string }
        autor_handle: { type: string }
        texto_original: { type: string }
        idioma: { type: string }
        publicado_en: { type: string, format: date-time }
        sentimiento: { type: string }
        score_sentimiento: { type: number }
        temas: { type: array, items: { type: string } }
    PaginatedOpiniones:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/OpinionOut" }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
    SentimentSeries:
      type: object
      properties:
        granularidad: { type: string }
        puntos:
          type: array
          items:
            type: object
            properties:
              timestamp: { type: string, format: date-time }
              positivo: { type: integer }
              neutro: { type: integer }
              negativo: { type: integer }
              score_medio: { type: number }

    # ---- CMS ----
...
    ChatQueryIn:
      type: object
      required: [sesion_id, canal, pregunta]
      properties:
        sesion_id: { type: string }
        canal: { type: string, enum: [web, app, totem] }
        idioma: { type: string, enum: [es, en, de, fr], default: es }
        pregunta: { type: string }
        contexto: { type: object }
    ChatResponseOut:
      type: object
      properties:
        respuesta: { type: string }
        intent_detectado: { type: string }
        nivel_confianza: { type: string, enum: [alta, media, fuera_de_dominio] }
        fuentes:
          type: array
          items:
            type: object
            properties:
              tipo: { type: string }
              referencia: { type: string }
              fecha: { type: string, format: date }
        sugerencias:
          type: array
          items: { type: string }
...
    MonthlyReport:
      type: object
      properties:
        periodo:
          type: object
          properties:
            year: { type: integer }
            month: { type: integer }
        disponibilidad_por_componente:
          type: object
          additionalProperties: { type: number }
        ans:
          type: object
          properties:
            criticas:
              type: object
              properties:
                tickets: { type: integer }
                tiempo_medio_respuesta_h: { type: number }
                tiempo_medio_resolucion_h: { type: number }
                cumplimiento_porc: { type: number }
            altas: { $ref: "#/components/schemas/AnsBucket" }
            medias: { $ref: "#/components/schemas/AnsBucket" }
            bajas: { $ref: "#/components/schemas/AnsBucket" }
        kpis_uso:
          type: object
          properties:
            interacciones_totems: { type: integer }
            sesiones_chatbot: { type: integer }
            visitas_web: { type: integer }
            sesiones_app: { type: integer }
        seguridad:
          type: object
          properties:
            eventos_waf_bloqueados: { type: integer }
            alertas_edr: { type: integer }
            incidentes_confirmados: { type: integer }
            brechas_notificadas_aepd: { type: integer }
        acciones_preventivas:
          type: array
          items:
            type: object
            properties:
              accion: { type: string }
              ejecutada: { type: boolean }
              fecha: { type: string, format: date }
    AnsBucket:
      type: object
      properties:
        tickets: { type: integer }
        tiempo_medio_respuesta_h: { type: number }
        tiempo_medio_resolucion_h: { type: number }
        cumplimiento_porc: { type: number }